refactor: split type definitions to header file
